+++ intern/vray_for_blender_rt/src/vfb_export_settings.h
@@ -66,7 +66,7 @@ struct SettingsAnimation {
 	{}
 
 	int           use;
-	int           frame_current;
+	float         frame_current;
 	int           frame_start;
 	int           frame_step;
 	AnimationMode mode;
